+++ di/src/test/java/com/interface21/beans/factory/support/ConfigurationBeanScannerTest.java
@@ -11,14 +11,14 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-class ConfigurationScannerTest {
+class ConfigurationBeanScannerTest {
 
     @Test
     @DisplayName("Configuration 클래스의 @ComponentScan 을 통해 basePackage 정보들을 조회할 수 있다.")
     void getBasePackagesTest() {
-        final ConfigurationScanner configurationScanner = new ConfigurationScanner(List.of(ComponentScanWithDefault.class, ComponentScanWithValue.class, ComponentScanWithBasePackages.class));
+        final ConfigurationBeanScanner configurationBeanScanner = new ConfigurationBeanScanner(List.of(ComponentScanWithDefault.class, ComponentScanWithValue.class, ComponentScanWithBasePackages.class));
 
-        final Object[] basePackages = configurationScanner.getBasePackages();
+        final Object[] basePackages = configurationBeanScanner.getBasePackages();
 
         assertThat(basePackages).containsExactlyInAnyOrder(
                 "com.interface21.beans.factory.support",
@@ -31,16 +31,16 @@ class ConfigurationScannerTest {
     @Test
     @DisplayName("@Configuration 이 없는 클래스가 있으면 예외를 던진다..")
     void testEmptyBasePackage() {
-        assertThatThrownBy(() -> new ConfigurationScanner(List.of(ConfigWithoutAnnotationClass.class)))
+        assertThatThrownBy(() -> new ConfigurationBeanScanner(List.of(ConfigWithoutAnnotationClass.class)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     @DisplayName("Configuration 클래스의 @Bean 메서드를 찾아 BeanRegistry 에 등록한다")
     void scanTest() {
-        final ConfigurationScanner configurationScanner = new ConfigurationScanner(List.of(TestConfig.class));
+        final ConfigurationBeanScanner configurationBeanScanner = new ConfigurationBeanScanner(List.of(TestConfig.class));
 
-        final BeanDefinitionRegistry beanDefinitionRegistry = configurationScanner.scan();
+        final BeanDefinitionRegistry beanDefinitionRegistry = configurationBeanScanner.scan();
 
         assertThat(beanDefinitionRegistry.getBeanDefinition(TestBean.class)).isNotNull();
